split up `ledger_entry` and `ledger` tests
